requirements.txt
langchain
langchain-core
langchain-google-genai
google-generativeai
pypdf
faiss-cpu
numpy
google-search-results
streamlit
pandas
openpyxl
plotly
python-dotenv
langchain-openai
openai